pyproject.toml
[project]
name = "briefler"
version = "0.1.0"
description = "briefler using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.2.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0"
]

[project.scripts]
kickoff = "briefler.main:kickoff"
run_crew = "briefler.main:kickoff"
plot = "briefler.main:plot"
run_with_trigger = "briefler.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/api", "src/briefler"]

[tool.crewai]
type = "flow"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = [
    "--import-mode=importlib",
    "-ra",
    "--strict-markers",
]
